src/impl/util/types.ts
export type PrimitiveValue = string | number | boolean;

export function isPrimitive(value: any): boolean {
  return (typeof value !== 'object' && typeof value !== 'function') || value === null
}

export function unsafeCast<T>(x: any): T {
  return x as T;
}

export function typeDescription(x: any): string {
  if (x === null)
    return 'null';

  let t = typeof x;
  if (t !== 'object') return t;

  const p = Object.getPrototypeOf(x);
  if (p !== Object.prototype)
    return p.constructor.name;

  return t
}

export type Constructor<T = any> = new(...args: any[]) => T ;